% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{fbs_barplot}
\alias{fbs_barplot}
\title{FUNCTION_TITLE}
\usage{
fbs_barplot(
  data,
  aesthetic,
  error = NULL,
  title = NULL,
  value_name = NULL,
  horizontal = FALSE,
  label_bars = FALSE,
  label_position = "top",
  continuous_format = "comma"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{aesthetic}{PARAM_DESCRIPTION}

\item{error}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{value_name}{PARAM_DESCRIPTION, Default: NULL}

\item{horizontal}{PARAM_DESCRIPTION, Default: FALSE}

\item{label_bars}{PARAM_DESCRIPTION, Default: FALSE}

\item{label_position}{PARAM_DESCRIPTION, Default: 'top'}

\item{continuous_format}{PARAM_DESCRIPTION, Default: 'comma'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{aes}}
}
