% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{fbs_stackplot}
\alias{fbs_stackplot}
\title{FUNCTION_TITLE}
\usage{
fbs_stackplot(
  data,
  aesthetic,
  title = NULL,
  value_name = NULL,
  horizontal = FALSE,
  continuous_format = "comma"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{aesthetic}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{value_name}{PARAM_DESCRIPTION, Default: NULL}

\item{horizontal}{PARAM_DESCRIPTION, Default: FALSE}

\item{continuous_format}{PARAM_DESCRIPTION, Default: 'comma'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{aes}}
}
