% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{fbs_templates}
\alias{fbs_templates}
\alias{fbs_barplot}
\alias{fbs_stackplot}
\alias{fbs_distribution_plot}
\alias{fbs_lineplot}
\title{Plot Templates}
\usage{
fbs_barplot(
  data,
  aesthetic,
  error = NULL,
  title = NULL,
  value_name = NULL,
  horizontal = FALSE,
  legend_hide = FALSE,
  continuous_format = "comma",
  font = "sans",
  text_scale = 1
)

fbs_stackplot(
  data,
  aesthetic,
  title = NULL,
  value_name = NULL,
  horizontal = FALSE,
  legend_hide = FALSE,
  continuous_format = "comma",
  font = "sans",
  text_scale = 1
)

fbs_distribution_plot(
  data,
  aesthetic,
  title = NULL,
  value_name = NULL,
  horizontal = FALSE,
  legend_hide = FALSE,
  continuous_format = "percent",
  font = "sans",
  text_scale = 1
)

fbs_lineplot(
  data,
  aesthetic,
  title = NULL,
  value_name = NULL,
  legend_hide = FALSE,
  continuous_format = "comma",
  font = "sans",
  text_scale = 1
)
}
\arguments{
\item{data}{Default dataset used for plot creation, passed into \code{\link[ggplot2]{ggplot}}.}

\item{aesthetic}{Aesthetic mapping used for the plot created by \link[ggplot2]{aes}. See details
for a list of accepted aesthetics for each template.}

\item{error}{Column containing values to display as error bars within \code{fbs_barplot()}. Currently
only supports symmetrical errors. Default: NULL}

\item{title}{Character string of text used as the plot title. Default: NULL}

\item{value_name}{Character string of text used to label the continuous axis, Default: NULL}

\item{horizontal}{Set to \code{TRUE} to change the chart orientation to horizontal,
Default: FALSE}

\item{legend_hide}{Set to \code{TRUE} to hide the plot legend, Default: FALSE}

\item{continuous_format}{Character string denoting the format that the continuous (y) axis should
take.

One of "identity", "comma", "percent", "abbreviated" or "scientific". Defaults to "percent" for
\code{fbs_distribution_plot()} and "comma" for all other plots.}

\item{font}{Change the default font.}

\item{text_scale}{Scale up or down the text size within the plot.}
}
\value{
Returns a ggplot object
}
\description{
Set of functions containing templates for creating publication ready charts in Defra's
Farm Business Survey style.
}
\details{
This set of functions can be used in isolation to generate basic publication ready charts in the
FBS style or can be combined with other ggplot2 layers for a deep level of customisation.

For a list of accepted aesthetics for \code{fbs_barplot}, \code{fbs_stackplot} and
\code{fbs_distribution_plot}, see \code{\link[ggplot2]{geom_bar}}.

For a list of accepted aesthetics for \code{fbs_lineplot} see \code{\link[ggplot2]{geom_line}}.

The following aesthetics are automatically scaled if supplied with their respective
\code{scale_*_govuk} scale. All other aesthetics must be scaled manually with additional ggplot
layers.

\itemize{
  \item \code{fill: \link{scale_fill_govuk}}
  \item \code{colour: \link{scale_colour_govuk}}
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_bar}},
\code{\link[ggplot2]{geom_line}}
}
