% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_fbsbar}
\alias{geom_fbsbar}
\title{FUNCTION_TITLE}
\usage{
geom_fbsbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "dodge",
  ...,
  width = 0.75,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{stat}{PARAM_DESCRIPTION, Default: 'identity'}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge'}

\item{...}{PARAM_DESCRIPTION}

\item{width}{PARAM_DESCRIPTION, Default: 0.75}

\item{na.rm}{PARAM_DESCRIPTION, Default: FALSE}

\item{orientation}{PARAM_DESCRIPTION, Default: NA}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
