% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_fbsline}
\alias{geom_fbsline}
\title{FUNCTION_TITLE}
\usage{
geom_fbsline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{stat}{PARAM_DESCRIPTION, Default: 'identity'}

\item{position}{PARAM_DESCRIPTION, Default: 'identity'}

\item{na.rm}{PARAM_DESCRIPTION, Default: FALSE}

\item{orientation}{PARAM_DESCRIPTION, Default: NA}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{layer}}
}
