% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ggfbs_colours}
\alias{ggfbs_colours}
\alias{gss_cols}
\alias{govuk_cols}
\title{Colour functions}
\usage{
gss_cols(...)

govuk_cols(...)
}
\arguments{
\item{...}{character or numeric vector denoting the colours of the palette to return.}
}
\description{
Generate a named vector of colours in the order they would appear within the
automatically generated charts.
}
\details{
Default colour palette: gss.
}
\examples{
# Return all colours in the palette
gss_cols()

# Return the first colour in the palette by number
gss_cols(1)

# Return multiple colours in palette by number
gss_cols(1:3)
gss_cols(3:6)

# Return colours by name
gss_cols("dark-blue")
gss_cols(c("dark-blue", "turquoise", "dark-pink", "orange"))
}
