% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{govuk_pal}
\alias{govuk_pal}
\title{FUNCTION_TITLE}
\usage{
govuk_pal()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
