% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pallettes}
\alias{pallettes}
\alias{gss_pal}
\alias{govuk_pal}
\title{Colour Palettes}
\usage{
gss_pal()

govuk_pal()
}
\description{
Palette functions similar to those found within the scales package. (e.g.
\code{\link[scales]{hue_pal}}).
}
\details{
Default colour palate: gss.


Calling these functions will return the palette function passed into
\code{\link{scale_colour_govuk}} and \code{\link[=scale_colour_govuk]{scale_fill_govuk}}.

This function can also be used to return a character vector containing \code{x} number of
colours with the syntax \code{gss_pal()(x)}. For example, \code{gss_pal()(5)} will return a
character vector of 5 colours. \strong{Note}: Unlike \code{\link{gss_cols}}, named colours are
not accepted. Only a single number may be passed into the palette.
}
\examples{
# Return the palette function
gss_pal()
govuk_pal()

# Return a vector of colours
gss_pal()(4)
govuk_pal()(4)
}
