% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_govuk}
\alias{scale_color_govuk}
\title{FUNCTION_TITLE}
\usage{
scale_color_govuk(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
