% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fbs}
\alias{theme_fbs}
\title{Farm Business Survey Theme}
\usage{
theme_fbs(horizontal = FALSE, text_scale = 1, font = "sans")
}
\arguments{
\item{horizontal}{Set to \code{TRUE} to adjust the theme for horizontal plots.}

\item{text_scale}{Scale up or down the text size within the plot.}

\item{font}{Change the default font.}
}
\description{
Apply the Farm Business Survey (FBS) theme to any plot
}
\details{
Applies the default FBS theme to a plot. By default this is applied to any of the chart templates
within this package.

This function enables the application of the FBS theme to charts which do not have a set
template. This can be combined with the scale_*_govuk set of functions to match the theming and
colour scheme used within the templates.
}
\examples{
# Add the FBS theme to a basic boxplot
ggplot(mpg, aes(class, hwy, colour = drv)) +
  geom_boxplot(size = 1.2) +
  scale_colour_govuk() +
  theme_fbs()
}
\seealso{
\code{\link[ggplot2]{theme}}
}
