% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{govuk_cols}
\alias{govuk_cols}
\title{FUNCTION_TITLE}
\usage{
govuk_cols(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
