% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_govuk}
\alias{scale_colour_govuk}
\alias{scale_fill_govuk}
\title{Discrete scales in the gov.uk style}
\usage{
scale_colour_govuk(...)

scale_fill_govuk(...)
}
\arguments{
\item{...}{Other arguments passed on to \link[ggplot2]{discrete_scale}}
}
\description{
Provides a discrete colour scheme in the gov.uk type. See \code{\link{govuk_cols}} for the list
of colours used within these scales.
}
\details{
These scales are automatically applied within each chart template when their respective
aesthetics is provided within the aesthetic specification.

These scales can additionally be added to non-template charts to provide the colour scheme used
within the templates.
}
\examples{
\dontrun{
# Add the FBS theme to a basic boxplot
ggplot(mpg, aes(class, hwy, colour = drv)) +
  geom_boxplot(size = 1.2) +
  scale_colour_govuk() +
  theme_fbs()
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
